fix: 修复循环引用导致单测错误
